Add form for creating new foods
This is achieved by using a previous food as a template. This probably
does not benefit the foods workflow much since one is likely to change
all values anyway, but it should not hurt it either for the same reason.
Additionally this is the flow that will be used to create meals where
one meal is likely to share a lot of components with some previous meal
and consistency makes it easier to understand the app. One field is not
automatically copied over in order to make it harder creating identical
copies accidentally.